+++ web/include/class/browser/PhantomJSWrapper.php
@@ -13,6 +13,20 @@ abstract class PhantomJSWrapper {
     protected $_aHeaders = array();
 
     /**
+     * Request specific arguments. (vs client arguments)
+     * Should be referred when get() is performed.
+     * @remark  should be overridden by an extended class.
+     * @var array
+     */
+    protected $_aRequestArguments = array();
+    private   $___aDefaultRequestArguments = array(
+        'method'    => 'GET',
+        'timeout'   => 10000,   // n seconds * 1000
+        'delay'     => 5,       // delay to wait for complete page load
+        'data'      => array(),  // for POST data
+    );
+
+    /**
      * PhantomJSWrapper constructor.
      *
      * @param string string          $sPhantomJSBinaryPath
@@ -37,7 +51,11 @@ abstract class PhantomJSWrapper {
         $this->_sUserAgent = $sUserAgent;
         $this->_aHeaders   = $aHeaders;
 
+        // Format default request arguments.
+        $this->_aRequestArguments = $this->_aRequestArguments + $this->___aDefaultRequestArguments;
+
     }
+
         /**
          * @param Client $oClient
          * @param $asConfig
@@ -79,4 +97,16 @@ abstract class PhantomJSWrapper {
                 }
                 return ( string ) $mValue;
             }
+
+    /**
+     * @return array
+     */
+    protected function _getRequestArguments( array $aOverride=array() ) {
+        return $aOverride + $this->_aRequestArguments;
+    }
+    public function setRequestArguments( array $aOverride=array() ) {
+        $this->_aRequestArguments = $aOverride + $this->_aRequestArguments;
+    }
+
+    // public function get(  ) { return ''; }
 }